requirements.txt
django
djangorestframework
Pillow
drf-yasg
